Move contact email verification to pontoon.messaging and use HTML template
