+++ pontoon/contributors/tests/test_utils.py
@@ -6,7 +6,6 @@ import pytest
 
 from dateutil.relativedelta import relativedelta
 
-from django.test.client import RequestFactory
 from django.utils import timezone
 
 from pontoon.actionlog.models import ActionLog
@@ -108,22 +107,6 @@ def test_generate_verification_token(member):
         args = mock_encode.call_args.args
         assert list(args[0].values())[0] == member.user.pk
         assert list(args[0].values())[1] == member.user.profile.contact_email
-
-
-@pytest.mark.django_db
-def test_send_verification_email(member):
-    with patch("pontoon.contributors.utils.EmailMessage") as mock_email_message:
-        rf = RequestFactory()
-        request = rf.get("/settings/")
-        request.user = member.user
-
-        token = "EMAIL-VERIFICATION-TOKEN"
-        utils.send_verification_email(request, token)
-        assert mock_email_message.called
-
-        kwargs = mock_email_message.call_args.kwargs
-        assert token in kwargs["body"]
-        assert kwargs["to"] == [None]
 
 
 @pytest.mark.django_db
